Remove src path
