+++ assets/less/site.less
@@ -1,7 +1,7 @@
-@import 'base/_settings.less';
-@import 'base/_mixins.less';
-@import 'base/reset.less';
-@import 'base/syntax.less';
+@import '_settings.less';
+@import '_mixins.less';
+@import 'reset.less';
+@import 'syntax.less';
 
 *, ::before, ::after {
     -moz-box-sizing: border-box;
