--- src/readme.md
@@ -1,23 +1,0 @@
-Customizing Vault
------------------
-
-Requirements for building
-- [Node.js](https://nodejs.org/en/)
-- [Yarn](https://yarnpkg.com/en/)
-
-Install dependencies:
-```
-yarn install
-```
-
-To build run the command -
-
-```
-yarn run build
-```
-
-If you want a non-minimized stylesheet you can run the following command
-
-```
-yarn run builddev
-```
